Rename CXC to CxC across codebase
Updated all references from 'CXC' to 'CxC' for branding consistency in metadata, UI text, documentation, and code comments. Refactored component names and file paths to match the new naming convention. Removed unused theme-toggle components from both cxc and web apps.
