--- apps/web/components/theme-toggle.tsx
@@ -1,24 +1,0 @@
-"use client";
-
-import * as React from "react";
-import { Moon, Sun } from "lucide-react";
-import { useTheme } from "next-themes";
-
-import { Button } from "@uwdsc/ui";
-
-export function ThemeToggle() {
-  const { setTheme, theme } = useTheme();
-
-  return (
-    <Button
-      variant="outline"
-      size="icon"
-      className="h-10 w-10"
-      onClick={() => setTheme(theme === "light" ? "dark" : "light")}
-    >
-      <Sun className="h-[1.4rem] w-[1.4rem] rotate-0 scale-100 transition-all dark:-rotate-90 dark:scale-0" />
-      <Moon className="absolute h-[1.4rem] w-[1.4rem] rotate-90 scale-0 transition-all dark:rotate-0 dark:scale-100" />
-      <span className="sr-only">Toggle theme</span>
-    </Button>
-  );
-}
